requirements.txt
Django==1.11
django-cors-headers==2.4.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
mysqlclient==1.3.13
Pillow==5.2.0
PyJWT==1.6.4
pytz==2018.5
